Simplify State and Trajectory
