+++ chirho/dynamical/internals/indexed.py
@@ -1,5 +1,7 @@
 from typing import TypeVar
 
+import torch
+
 from chirho.dynamical.ops.dynamical import State, Trajectory
 from chirho.indexed.ops import IndexSet, gather, indices_of, union
 
@@ -51,3 +53,28 @@ def _gather_trajectory(
             for k in trj.keys
         }
     )
+
+
+def _index_last_dim_with_mask(x: torch.Tensor, mask: torch.Tensor) -> torch.Tensor:
+    # Index into the last dimension of x with a boolean mask.
+    # TODO AZ — There must be an easier way to do this?
+    # NOTE AZ — this could be easily modified to support the last n dimensions, adapt if needed.
+
+    if mask.dtype != torch.bool:
+        raise ValueError(
+            f"_index_last_dim_with_mask only supports boolean mask indexing, but got dtype {mask.dtype}."
+        )
+
+    # Require that the mask is 1d and aligns with the last dimension of x.
+    if mask.ndim != 1 or mask.shape[0] != x.shape[-1]:
+        raise ValueError(
+            "_index_last_dim_with_mask only supports 1d boolean mask indexing, and must align with the last "
+            f"dimension of x, but got mask shape {mask.shape} and x shape {x.shape}."
+        )
+
+    return torch.masked_select(
+        x,
+        # Get a shape that will broadcast to the shape of x. This will be [1, ..., len(mask)].
+        mask.reshape((1,) * (x.ndim - 1) + mask.shape)
+        # masked_select flattens tensors, so we need to reshape back to the original shape w/ the mask applied.
+    ).reshape(x.shape[:-1] + (int(mask.sum()),))
